#579 Split simpleworker into a connectedworker.
